+++ mortar-sample/src/main/java/com/example/mortar/DemoApplication.java
@@ -16,10 +16,16 @@
 package com.example.mortar;
 
 import android.app.Application;
+import com.example.mortar.core.ApplicationModule;
+import dagger.ObjectGraph;
+import mortar.Mortar;
+import mortar.MortarScope;
 import rx.plugins.RxJavaErrorHandler;
 import rx.plugins.RxJavaPlugins;
 
 public class DemoApplication extends Application {
+  private MortarScope rootScope;
+
   @Override public void onCreate() {
     super.onCreate();
 
@@ -29,5 +35,12 @@ public class DemoApplication extends Application {
         throw new RuntimeException(e);
       }
     });
+
+    rootScope =
+        Mortar.createRootScope(BuildConfig.DEBUG, ObjectGraph.create(new ApplicationModule()));
+  }
+
+  public MortarScope getRootScope() {
+    return rootScope;
   }
 }